requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
cachetools==4.2.4
charset-normalizer==3.1.0
frozenlist==1.3.3
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.59.0
grpcio==1.53.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
openai==0.27.0
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==1.0.0
python-telegram-bot==20.2
python-telegram-bot-raw==20.2
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
telegram==0.0.1
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
pymongo~=4.3.3